@@ -20,8 +20,6 @@
   "dependencies": {
     "dts-bundle-generator": "^6.9.0",
     "react-quill": "^1.3.5",
-    "ts": "^0.2.2",
-    "ts-node": "^10.7.0",
     "types-webpack-bundler": "^1.0.2",
     "web-vitals": "^1.0.1"
   },
@@ -31,7 +29,7 @@
     "package": "rimraf /F /Q pkg_build && babel src/lib --out-dir pkg_build --copy-files && yarn pack",
     "bundle": "rimraf /F /Q pkg_build && webpack --mode production --config webpack.config.js --entry /src/lib/index.ts -o pkg_build",
     "test-api": "ts-node testing-api-server/test-api.ts",
-    "test": "jest",
+    "test": "node --experimental-vm-modules  node_modules/jest/bin/jest.js --config jest.config.js",
     "test-bundle": "webpack --config ./webpack.config.test.js",
     "deprecated-types": "tsc --project types.tsconfig.json --emitDeclarationOnly"
   },
@@ -59,29 +57,37 @@
   },
   "devDependencies": {
     "@babel/cli": "^7.16.0",
-    "@babel/core": "^7.15.8",
+    "@babel/core": "^7.18.2",
     "@babel/preset-env": "^7.15.8",
     "@babel/preset-react": "^7.14.5",
+    "@babel/preset-typescript": "^7.17.12",
     "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/preact": "^2.0.1",
+    "@testing-library/preact": "^3.2.0",
     "@testing-library/user-event": "^12.1.10",
-    "@types/node": "^17.0.35",
-    "babel-jest": "^27.3.1",
+    "@types/jest": "^27.5.1",
+    "@types/node": "^17.0.36",
+    "babel-jest": "^28.1.0",
     "babel-loader": "^8.2.3",
     "buffer": "^6.0.3",
     "css-loader": "^6.5.0",
     "declaration-bundler-webpack-plugin": "^1.0.3",
     "express": "^4.18.1",
     "identity-obj-proxy": "^3.0.0",
-    "jest": "^27.3.1",
+    "jest": "^28.1.0",
+    "jest-environment-jsdom": "^28.1.0",
+    "jest-preset-preact": "^4.0.5",
+    "jsdom": "^19.0.0",
     "preact": "^10.7.2",
     "preact-render-to-string": "^5.1.19",
     "rimraf": "^3.0.2",
     "sass": "^1.43.2",
     "sass-loader": "^12.3.0",
     "style-loader": "^3.3.1",
+    "ts": "^0.2.2",
+    "ts-jest": "^28.0.3",
     "ts-loader": "^9.2.9",
-    "typescript": "^4.6.4",
+    "ts-node": "^10.8.0",
+    "typescript": "^4.7.2",
     "webpack": "^5.60.0",
     "webpack-cli": "^4.9.1"
   }
